OR13: When there is a server error, user gets a message and ac tion is aborted. Also fixed: not closing tabs when unit is deleted.

+++ packages/core/src/storage/InMemoryModel.ts
@@ -1,43 +1,63 @@
-import { autorun, makeObservable, observable, runInAction } from 'mobx';
-import type {FreModel, FreModelUnit} from "../ast/index.js";
+import { autorun, makeObservable, observable, runInAction } from "mobx"
+import type { FreModel, FreModelUnit } from "../ast/index.js"
 import {
     AST,
     FreChangeManager,
-    FrePartDelta,
-    FrePartListDelta,
-    FrePrimDelta,
-    FrePrimListDelta, FreUndoManager
-} from '../change-manager/index.js';
-import type {FreEnvironment} from "../environment/index.js";
-import {FreLogger} from "../logging/index.js";
-import {isNullOrUndefined, notNullOrUndefined} from "../util/index.js";
-import type {IServerCommunication, FreUnitIdentifier} from "./server/index.js";
+    type FrePartDelta,
+    type FrePartListDelta,
+    type FrePrimDelta,
+    type FrePrimListDelta,
+    FreUndoManager,
+} from "../change-manager/index.js"
+import type { FreEnvironment } from "../environment/index.js"
+import { FreLogger } from "../logging/index.js"
+import { isNullOrUndefined, notNullOrUndefined } from "../util/index.js"
+import { FreErrorSeverity } from "../validator/index.js"
+import type { FreUnitIdentifier, IServerCommunication } from "./server/index.js"
 
-export type ModelChangedCallbackFunction = (m: InMemoryModel) => void;
+export type ModelChangedCallbackFunction = (m: InMemoryModel) => void
 
-const LOGGER: FreLogger = new FreLogger("InMemoryModel").mute();
+export class InMemoryError {
+    message: string
+    constructor (msg: string) {
+        this.message = msg
+    }
+}
+export function isInMemoryError(object: unknown): object is InMemoryError {
+    return object instanceof InMemoryError
+}
+
+const LOGGER: FreLogger = new FreLogger("InMemoryModel").mute()
 
 export class InMemoryModel {
-    private languageEnvironment: FreEnvironment;
-    private server: IServerCommunication;
-    model: FreModel | undefined = undefined;
+    private languageEnvironment: FreEnvironment
+    private server: IServerCommunication
+    model: FreModel | undefined = undefined
     // units that have been changed but not saved
-    dirtyUnits: Set<FreModelUnit> = new Set();
+    dirtyUnits: Set<FreModelUnit> = new Set()
 
     constructor(languageEnvironment: FreEnvironment, server: IServerCommunication) {
-        this.languageEnvironment = languageEnvironment;
-        this.server = server;
-        makeObservable(this, { model: observable });
-        autorun( () => {
+        if (languageEnvironment === undefined) {
+            console.trace()
+        }
+        this.languageEnvironment = languageEnvironment
+        this.server = server
+        makeObservable(this, { model: observable })
+        autorun(() => {
             if (notNullOrUndefined(this.model)) {
                 this.model.getUnits()
                 this.currentModelChanged()
             }
-        });
-        FreChangeManager.getInstance().subscribeToPart(this.partChanged);
-        FreChangeManager.getInstance().subscribeToPrimitive(this.primChanged);
-        FreChangeManager.getInstance().subscribeToList(this.listChanged);
-        FreChangeManager.getInstance().subscribeToListElement(this.listElementChanged);
+        })
+        FreChangeManager.getInstance().subscribeToPart(this.partChanged)
+        FreChangeManager.getInstance().subscribeToPrimitive(this.primChanged)
+        FreChangeManager.getInstance().subscribeToList(this.listChanged)
+        FreChangeManager.getInstance().subscribeToListElement(this.listElementChanged)
+    }
+
+    onInMemoryError = (msg: string, severity?: FreErrorSeverity): void => {
+        console.error("onInMemoryError: " + msg + severity?.toString())
+        this.languageEnvironment?.editor.setUserMessage(msg, severity)
     }
 
     /**
@@ -45,26 +65,33 @@ export class InMemoryModel {
      * After this call the newly created model can be retrieved using _getModel_.
      * @param name
      */
-    async createModel(name: string): Promise<FreModel> {
+    async createModel(name: string): Promise<FreModel | InMemoryError> {
         LOGGER.log(`createModel ${name}`)
         runInAction(() => {
-            this.model = this.languageEnvironment.newModel(name);
+            this.model = this.languageEnvironment.newModel(name)
         })
-        await this.server.createModel(name);
-        FreUndoManager.getInstance().cleanAllStacks();
-        return this.model;
+        const response = await this.server.createModel(name)
+        FreUndoManager.getInstance().cleanAllStacks()
+        if (response.errors.length !== 0) {
+            this.onInMemoryError(response.errors[0], FreErrorSeverity.Info)
+            return new InMemoryError(response.errors[0])
+        }
+        return this.model
     }
 
     /**
      * Delete current model from the server.
      * After this call the current model is undefined.
      */
-    async deleteModel(): Promise<void> {
-        await this.server.deleteModel(this.model.name);
-        runInAction( () => {
-            this.model = undefined;
-        });
-        FreUndoManager.getInstance().cleanAllStacks();
+    async deleteModel(): Promise<void | InMemoryError> {
+        const response = await this.server.deleteModel(this.model.name)
+        if (response.errors.length > 0) {
+            return new InMemoryError(response.errors[0])
+        }
+        runInAction(() => {
+            this.model = undefined
+        })
+        FreUndoManager.getInstance().cleanAllStacks()
     }
 
     /**
@@ -72,39 +99,56 @@ export class InMemoryModel {
      * After this call the newly opened model can be retrieved using _getModel_.
      * * @param name
      */
-    async openModel(name: string): Promise<FreModel> {
-        LOGGER.log("openModel(" + name + ")");
-        AST.change( () => {
-            this.model = this.languageEnvironment.newModel(name);
+    async openModel(name: string): Promise<FreModel | InMemoryError> {
+        LOGGER.log("openModel(" + name + ")")
+        AST.change(() => {
+            this.model = this.languageEnvironment.newModel(name)
         })
-        const unitsIds = await this.server.loadUnitList(name);
-        for (const unitId of unitsIds) {
-            LOGGER.log("openModel: load model-unit: " + unitId.name);
-            const unit = await this.server.loadModelUnit(this.model.name, unitId);
-            AST.change( () => {
-                this.model.addUnit(unit as FreModelUnit);
+        const response = await this.server.loadUnitList(name)
+        if  (response.errors.length > 0) {
+            this.onInMemoryError(response.errors[0])
+            return new InMemoryError(response.errors[0])
+        }
+        for (const unitId of response.result) {
+            LOGGER.log("openModel: load model-unit: " + unitId.name)
+            const unit = await this.server.loadModelUnit(this.model.name, unitId)
+            AST.change(() => {
+                this.model.addUnit(unit.result as FreModelUnit)
             })
         }
-        FreUndoManager.getInstance().cleanAllStacks();
-        return this.model;
+        FreUndoManager.getInstance().cleanAllStacks()
+        return this.model
     }
 
-    async saveModel() {
-        LOGGER.log('InMemoryModel.saveModel()');
+    async saveModel(): Promise<void | InMemoryError> {
+        LOGGER.log("InMemoryModel.saveModel()")
         // save all units that are 'dirty', i.e. that have been changed after the previous save
-        await this.dirtyUnits.forEach(unit => {
-            this.saveUnit(unit);
-        })
+        const savedUnits =  []
+        for (const unit of this.dirtyUnits) {
+            const response = await this.saveUnit(unit)
+            if (isInMemoryError(response)) {
+                // Clean dirty for units saved so far
+                savedUnits.forEach(saved => this.dirtyUnits.delete(saved))
+                return response
+            } else {
+                savedUnits.push(unit)
+            }
+        }
         // when done, clean 'dirtyUnits' prop
-        this.dirtyUnits.clear();
+        this.dirtyUnits.clear()
     }
 
     /**
      * Get a list of all model names that are available on the server.
      */
-    async getModels(): Promise<string[]> {
-        LOGGER.log(`getModels`);
-        return await this.server.loadModelList();
+    async getModels(): Promise<string[] | InMemoryError> {
+        LOGGER.log(`getModels`)
+        const response = await this.server.loadModelList()
+        if (response.errors.length > 0) {
+            this.onInMemoryError(response.errors[0], FreErrorSeverity.Info)
+            return new InMemoryError(response.errors[0])
+        }
+        return response.result 
     }
 
     /**
@@ -113,36 +157,47 @@ export class InMemoryModel {
      * @param name
      * @param unitConcept
      */
-    async createUnit(name: string, unitConcept: string): Promise<FreModelUnit> {
+    async createUnit(name: string, unitConcept: string): Promise<FreModelUnit | InMemoryError> {
         LOGGER.log(`createUnit ${name} of concept ${unitConcept}`)
-        const newUnit = this.model.newUnit(unitConcept);
-        runInAction( () => {
-            newUnit.name = name;
+        const newUnit = this.model.newUnit(unitConcept)
+        runInAction(() => {
+            newUnit.name = name
         })
-        await this.server.createModelUnit(this.model.name, newUnit);
-        return newUnit;
+        const response = await this.server.createModelUnit(this.model.name, newUnit)
+        if (response.errors.length > 0) {
+            return new InMemoryError(response.errors[0])
+        }
+        return newUnit
     }
 
     /**
      * Delete _unit_ from the model.
      * @param unit
      */
-    async deleteUnit(unit: FreModelUnit) {
-        await this.server.deleteModelUnit(this.model.name, { name: unit.name, id: unit.freId(), type: unit.freLanguageConcept() });
-        AST.change( () => {
-            this.model.removeUnit(unit);
+    async deleteUnit(unit: FreModelUnit): Promise<void | InMemoryError> {
+        const response = await this.server.deleteModelUnit(this.model.name, { name: unit.name, id: unit.freId(), type: unit.freLanguageConcept() })
+        if (response.errors.length > 0) {
+            this.onInMemoryError(response.errors[0])
+            return new InMemoryError(response.errors[0])
+        }
+        AST.change(() => {
+            this.model.removeUnit(unit)
         })
     }
 
     /**
      * Delete _unit_ from the model.
-     * @param unit
+     * @param unitId
      */
-    async deleteUnitById(unitId: FreUnitIdentifier) {
-        await this.server.deleteModelUnit(this.model.name, unitId);
-        const unit: FreModelUnit = this.getUnitById(unitId);
-        AST.change( () => {
-            this.model.removeUnit(unit);
+    async deleteUnitById(unitId: FreUnitIdentifier): Promise<void | InMemoryError> {
+        const response = await this.server.deleteModelUnit(this.model.name, unitId)
+        if (response.errors.length > 0) {
+            this.onInMemoryError(response.errors[0])
+            return new InMemoryError(response.errors[0])
+        }
+        const unit: FreModelUnit = this.getUnitById(unitId)
+        AST.change(() => {
+            this.model.removeUnit(unit)
         })
     }
 
@@ -151,8 +206,8 @@ export class InMemoryModel {
      * @param name
      */
     getUnitByName(name: string) {
-        LOGGER.log(`getUnitByName`);
-        return this.model.findUnit(name);
+        LOGGER.log(`getUnitByName`)
+        return this.model.findUnit(name)
     }
 
     /**
@@ -162,25 +217,25 @@ export class InMemoryModel {
      */
     async addUnit(unit: FreModelUnit): Promise<void> {
         LOGGER.log(`addUnit ${unit?.name}`)
-        AST.change( () => {
-            this.model.addUnit(unit);
+        AST.change(() => {
+            this.model.addUnit(unit)
         })
-        await this.saveUnit(unit);
+        await this.saveUnit(unit)
     }
 
     /**
      * @param id
      */
     getUnitById(id: FreUnitIdentifier): FreModelUnit {
-        LOGGER.log(`getUnitById: ${id.name}`);
-        return this.model.findUnit(id.name);
+        LOGGER.log(`getUnitById: ${id.name}`)
+        return this.model.findUnit(id.name)
     }
 
     /**
      * Get all units of the current model.
      */
     getUnits(): FreModelUnit[] {
-        LOGGER.log(`getUnits`);
+        LOGGER.log(`getUnits`)
         const units = this?.model?.getUnits()
         if (isNullOrUndefined(units)) {
             return []
@@ -193,14 +248,14 @@ export class InMemoryModel {
      * Get all unit identifiers of the current model.
      */
     getUnitIdentifiers(): FreUnitIdentifier[] {
-        LOGGER.log(`getUnitIdentifiers`);
+        LOGGER.log(`getUnitIdentifiers`)
         const units = this.model?.getUnits()
         if (isNullOrUndefined(units)) {
             return []
         } else {
             return units.map((u) => {
-                return { name: u.name, id: u.freId(), type: u.freLanguageConcept() };
-            });
+                return { name: u.name, id: u.freId(), type: u.freLanguageConcept() }
+            })
         }
     }
 
@@ -209,15 +264,24 @@ export class InMemoryModel {
      * This is done only when there are unsaved changes.
      * @param unit
      */
-    async saveUnit(unit: FreModelUnit): Promise<void> {
-        LOGGER.log(`saveModelUnit`);
+    async saveUnit(unit: FreModelUnit): Promise<void | InMemoryError> {
+        LOGGER.log(`saveModelUnit`)
         if (this.dirtyUnits.has(unit)) {
-            await this.server.saveModelUnit(this.model.name, {
-                name: unit.name,
-                id: unit.freId(),
-                type: unit.freLanguageConcept()
-            }, unit);
-            this.dirtyUnits.delete(unit);
+            const serverResponse = await this.server.saveModelUnit(
+                this.model.name,
+                {
+                    name: unit.name,
+                    id: unit.freId(),
+                    type: unit.freLanguageConcept(),
+                },
+                unit,
+            )
+            if (serverResponse.errors.length === 0) {
+                this.dirtyUnits.delete(unit)
+            } else {
+                this.onInMemoryError(serverResponse.errors[0])
+                return new InMemoryError(`${serverResponse.errors[0]})`)
+            }
         }
     }
 
@@ -227,8 +291,8 @@ export class InMemoryModel {
      * @param unitId
      */
     async saveUnitById(unitId: FreUnitIdentifier) {
-        const unit: FreModelUnit = this.getUnitById(unitId);
-        this.saveUnit(unit);
+        const unit: FreModelUnit = this.getUnitById(unitId)
+        this.saveUnit(unit)
     }
 
     /************************************************************
@@ -237,26 +301,26 @@ export class InMemoryModel {
 
     primChanged = (delta: FrePrimDelta) => {
         if (this.getUnits().includes(delta.unit)) {
-            this.dirtyUnits.add(delta.unit);
+            this.dirtyUnits.add(delta.unit)
         }
     }
     partChanged = (delta: FrePartDelta) => {
         if (this.getUnits().includes(delta.unit)) {
-            this.dirtyUnits.add(delta.unit);
+            this.dirtyUnits.add(delta.unit)
         }
     }
     listElementChanged = (delta: FrePartDelta | FrePrimDelta) => {
         if (this.getUnits().includes(delta.unit)) {
-            this.dirtyUnits.add(delta.unit);
+            this.dirtyUnits.add(delta.unit)
         }
     }
     listChanged = (delta: FrePartListDelta | FrePrimListDelta) => {
         if (this.getUnits().includes(delta.unit)) {
-            this.dirtyUnits.add(delta.unit);
+            this.dirtyUnits.add(delta.unit)
         }
     }
     hasChanges(): boolean {
-        return this.dirtyUnits.size > 0;
+        return this.dirtyUnits.size > 0
     }
 
     /************************************************************
@@ -266,11 +330,11 @@ export class InMemoryModel {
     /**
      * Callbacks to inform listeners that the currentmodel/currentunit has changed.
      */
-    private currentModelListeners: ModelChangedCallbackFunction[] = [];
+    private currentModelListeners: ModelChangedCallbackFunction[] = []
     addCurrentModelListener(l: ModelChangedCallbackFunction): void {
-        this.currentModelListeners.push(l);
+        this.currentModelListeners.push(l)
     }
     currentModelChanged(): void {
-        this.currentModelListeners.forEach((l) => l(this));
+        this.currentModelListeners.forEach((l) => l(this))
     }
 }
